удаление и создание утки через бд, валидация создания и изменения утки, подключены allure
