+++ .gitignore
@@ -1,3 +1,4 @@
 .gradle/
 .idea/
-.build/
+build/
+target/
