+++ build.gradle
@@ -1,5 +1,6 @@
 plugins {
     id 'java'
+    id 'io.qameta.allure' version '2.9.6'
 }
 
 group 'org.example'
@@ -18,8 +19,14 @@ dependencies {
     testImplementation group: 'org.slf4j', name: 'slf4j-reload4j', version: '2.0.7'
     testCompileOnly group: 'org.projectlombok', name: 'lombok', version: '1.18.28'
     testAnnotationProcessor group: 'org.projectlombok', name: 'lombok', version: '1.18.28'
+    testImplementation group: 'com.h2database', name: 'h2', version: '1.4.196'
+    testImplementation group: 'com.consol.citrus', name: 'citrus-sql', version: '3.4.0'
 }
 
 test {
-    useTestNG()
+    useTestNG(){
+        parallel="classes"
+        threadCount 3
+        useDefaultListeners = true
+    }
 }
