+++ src/test/java/autotests/EndpointConfig.java
@@ -3,6 +3,7 @@ package autotests;
 import com.consol.citrus.http.client.HttpClient;
 import com.consol.citrus.http.client.HttpClientBuilder;
 import org.springframework.context.annotation.Bean;
+import org.springframework.jdbc.datasource.SingleConnectionDataSource;
 
 public class EndpointConfig {
 
@@ -11,5 +12,14 @@ public class EndpointConfig {
         return new HttpClientBuilder()
                 .requestUrl("http://localhost:2222")
                 .build();
+    }
+    @Bean("testDb")
+    public SingleConnectionDataSource db() {
+        SingleConnectionDataSource dataSource = new SingleConnectionDataSource();
+        dataSource.setDriverClassName("org.h2.Driver");
+        dataSource.setUrl("jdbc:h2:tcp://localhost:9092/mem:ducks");
+        dataSource.setUsername("dev");
+        dataSource.setPassword("dev");
+        return dataSource;
     }
 }
