+++ src/test/java/autotests/clients/DuckActionsClients.java
@@ -6,6 +6,7 @@ import com.consol.citrus.message.MessageType;
 import com.consol.citrus.message.builder.ObjectMappingPayloadBuilder;
 import com.consol.citrus.testng.spring.TestNGCitrusSpringSupport;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import io.qameta.allure.Step;
 import org.apache.http.entity.ContentType;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.ClassPathResource;
@@ -13,8 +14,11 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import com.consol.citrus.http.client.HttpClient;
 import com.consol.citrus.TestCaseRunner;
+import org.springframework.jdbc.datasource.SingleConnectionDataSource;
 import org.springframework.test.context.ContextConfiguration;
 
+import static com.consol.citrus.actions.ExecuteSQLAction.Builder.sql;
+import static com.consol.citrus.actions.ExecuteSQLQueryAction.Builder.query;
 import static com.consol.citrus.http.actions.HttpActionBuilder.http;
 import static com.consol.citrus.dsl.MessageSupport.MessageBodySupport.fromBody;
 
@@ -23,8 +27,25 @@ import static com.consol.citrus.dsl.MessageSupport.MessageBodySupport.fromBody;
 public class DuckActionsClients extends TestNGCitrusSpringSupport {
     @Autowired
     protected HttpClient yellowDuckService;
+    @Autowired
+    protected SingleConnectionDataSource testDb;
 
-    // Создание утки
+
+    public void validateDuck (TestCaseRunner runner, String id, String color, String height,
+                              String material, String sound, String wingsState) {
+        runner.$(query(testDb).statement("SELECT * FROM DUCK WHERE ID=" + id)
+                .validate("COLOR",color)
+                .validate("HEIGHT",height)
+                .validate("MATERIAL",material)
+                .validate("SOUND",sound)
+                .validate("WINGS_STATE",wingsState));
+    }
+
+    public void updateData (TestCaseRunner runner, String sql)
+    {
+        runner.$(sql(testDb).statement(sql));
+    }
+    @Step("Эндпоинт для создания уточки")
     public void create(TestCaseRunner runner, Object body) {
         runner.$(http().client(yellowDuckService)
                 .send()
@@ -34,7 +55,7 @@ public class DuckActionsClients extends TestNGCitrusSpringSupport {
                 .body(new ObjectMappingPayloadBuilder(body, new ObjectMapper())));
     }
 
-    // Удаление утки
+    @Step("Эндпоинт для удаления уточки")
     public void delete(TestCaseRunner runner) {
         runner.$(http().client(yellowDuckService)
                 .send()
@@ -43,7 +64,7 @@ public class DuckActionsClients extends TestNGCitrusSpringSupport {
                 .queryParam( "id", "${id}"));
     }
 
-    // Обновление свойств утки
+    @Step("Эндпоинт для обновления свойств уточки")
     public void update(TestCaseRunner runner, String color, double height, String material, String sound,
                        WingState wingsState) {
         runner.$(http().client(yellowDuckService)
@@ -59,7 +80,7 @@ public class DuckActionsClients extends TestNGCitrusSpringSupport {
          );
     }
 
-    // Получение свойств утки
+    @Step("Эндпоинт для получения свойств уточки")
     public void getProperties(TestCaseRunner runner) {
         runner.$(http().client(yellowDuckService)
                 .send()
@@ -68,7 +89,7 @@ public class DuckActionsClients extends TestNGCitrusSpringSupport {
                 .queryParam("id","${id}"));
     }
 
-    // Заставить утку летать
+    @Step("Эндпоинт заставить утку летать")
     public void fly(TestCaseRunner runner) {
         runner.$(http().client(yellowDuckService)
                 .send()
@@ -77,7 +98,7 @@ public class DuckActionsClients extends TestNGCitrusSpringSupport {
                 .queryParam("id", "${id}"));
     }
 
-    // Заставить утку плавать
+    @Step("Эндпоинт заставить утку плавать")
     public void swim(TestCaseRunner runner) {
         runner.$(http().client(yellowDuckService)
                 .send()
@@ -86,7 +107,7 @@ public class DuckActionsClients extends TestNGCitrusSpringSupport {
                 .queryParam("id", "${id}"));
     }
 
-    // Заставить утку крякать
+    @Step("Эндпоинт заставить утку крякать")
     public void quack(TestCaseRunner runner, String id, String repetitionCount, String soundCount) {
         runner.$(http().client(yellowDuckService)
                 .send()
